Create Longest Sub-Array with Sum K_gfg.cpp
Longest Sub-Array with Sum K
MediumAccuracy: 24.64%Submissions: 115K+Points: 4
Lamp
Stand out from the crowd. Prepare with Complete Interview Preparation  

Given an array containing N integers and an integer K., Your task is to find the length of the longest Sub-Array with the sum of the elements equal to the given value K.

 

Example 1:
 

Input :
A[] = {10, 5, 2, 7, 1, 9}
K = 15
Output : 4
Explanation:
The sub-array is {5, 2, 7, 1}.
Example 2:

Input : 
A[] = {-1, 2, 3}
K = 6
Output : 0
Explanation: 
There is no such sub-array with sum 6.
Your Task:
This is a function problem. The input is already taken care of by the driver code. You only need to complete the function lenOfLongSubarr() that takes an array (A), sizeOfArray (n),  sum (K)and returns the required length of the longest Sub-Array. The driver code takes care of the printing.

Expected Time Complexity: O(N).
Expected Auxiliary Space: O(N).

 

Constraints:
1<=N<=105
-105<=A[i], K<=105


https://practice.geeksforgeeks.org/problems/longest-sub-array-with-sum-k0809/1?page=1&status[]=unsolved&category[]=Hash&sortBy=submissions